añadidos los cambios de xavi
