+++ package.json
@@ -10,13 +10,15 @@
   },
   "dependencies": {
     "@mantine/core": "^4.0.7",
+    "@mantine/form": "^4.1.0",
     "@mantine/hooks": "^4.0.7",
     "@mantine/next": "^4.0.7",
     "leaflet": "^1.7.1",
     "next": "12.1.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-leaflet": "^3.2.5"
+    "react-leaflet": "^3.2.5",
+    "tabler-icons-react": "^1.43.0"
   },
   "devDependencies": {
     "@types/leaflet": "^1.7.9",
