+++ pages/admin/promociones/index.tsx
